requirements.txt
pip
python-dotenv